farmvet/portals/credentials.py
from decouple import config

MPESA_SHORTCODE = config("MPESA_SHORTCODE")  # Your Paybill or Till Number
MPESA_PASSKEY = config("MPESA_PASSKEY") 
CALLBACK_URL = config("CALLBACK_URL") 
MPESA_API_URL = config("MPESA_API_URL") 
# API Authentication
CONSUMER_KEY = config("CONSUMER_KEY") 
CONSUMER_SECRET =config("CONSUMER_SECRET")  
TILL_NO =config("TILL_NO")  
