farmvet/requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asttokens==2.2.1
backcall==0.2.0
certifi==2022.9.14
cffi==1.17.0
chardet==4.0.0
comm==0.1.2
cryptography==38.0.1
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.11
django-allauth==0.60.0
django-crispy-forms==1.9.2
django-htmx==1.17.3
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
entrypoints==0.4
executing==1.2.0
fonttools==4.37.4
fpdf2==2.5.7
future==0.18.2
html5lib==1.1
idna==2.10
ipykernel==6.20.2
ipython==8.8.0
jedi==0.18.2
jupyter-client==7.4.9
jupyter-core==5.1.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
oauthlib==3.2.1
packaging==23.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==2.6.2
postgres==4.0
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg2-binary==2.9.9
psycopg2-pool==1.2
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.4
python3-openid==3.2.0
pytz==2022.7.1
#pyzmq==25.0.0
#reportlab==3.6.13
requests==2.25.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
svg.path==6.2
tornado==6.2
traitlets==5.8.1
types-cryptography==3.3.23
typing-extensions==4.10.0
urllib3==1.26.14
wcwidth==0.2.6
webencodings==0.5.1
xhtml2pdf==0.2.11
